# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XogsideIQ - Crypto Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List or search tokens
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: limit
          in: query
          schema:
            type: integer
          description: Number of results
        - name: offset
          in: query
          schema:
            type: integer
          description: Offset for pagination
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
  /tokens/{symbol}:
    get:
      operationId: getToken
      tags: [tokens]
      summary: Get full token research data
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDetail"
        "404":
          description: Token not found
  /tokens/{symbol}/scores:
    get:
      operationId: getTokenScores
      tags: [tokens]
      summary: Get token intelligence scores and grade
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenScores"
  /tokens/{symbol}/ai-research:
    get:
      operationId: getTokenAiResearch
      tags: [tokens]
      summary: Get AI-generated research summary for a token
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI research summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResearch"
  /tokens/{symbol}/news:
    get:
      operationId: getTokenNews
      tags: [tokens]
      summary: Get news related to a token
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
...
components:
  schemas:
...
    Token:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        chain:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        priceChange24h:
          type: ["number", "null"]
        marketCap:
          type: ["number", "null"]
        volume24h:
          type: ["number", "null"]
        overallScore:
          type: ["integer", "null"]
        finalGrade:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, symbol, name, chain, createdAt]
    TokenDetail:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        chain:
          type: string
        description:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        whitepaperUrl:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        priceChange24h:
          type: ["number", "null"]
        priceChange7d:
          type: ["number", "null"]
        marketCap:
          type: ["number", "null"]
        volume24h:
          type: ["number", "null"]
        fdv:
          type: ["number", "null"]
        circulatingSupply:
          type: ["number", "null"]
        totalSupply:
          type: ["number", "null"]
        overallScore:
          type: ["integer", "null"]
        fundamentalScore:
          type: ["integer", "null"]
        technicalScore:
          type: ["integer", "null"]
        sentimentScore:
          type: ["integer", "null"]
        riskScore:
          type: ["integer", "null"]
        narrativeMomentumScore:
          type: ["integer", "null"]
        finalGrade:
          type: ["string", "null"]
        gradeExplanation:
          type: ["string", "null"]
        narratives:
          type: array
          items:
            $ref: "#/components/schemas/Narrative"
        createdAt:
          type: string
      required: [id, symbol, name, chain, createdAt, narratives]
    TokenScores:
      type: object
      properties:
        tokenId:
          type: integer
        symbol:
          type: string
        overallScore:
          type: integer
        fundamentalScore:
          type: integer
        technicalScore:
          type: integer
        sentimentScore:
          type: integer
        riskScore:
          type: integer
        narrativeMomentumScore:
          type: integer
        finalGrade:
          type: string
        gradeExplanation:
          type: string
        updatedAt:
          type: string
      required: [tokenId, symbol, overallScore, fundamentalScore, technicalScore, sentimentScore, riskScore, narrativeMomentumScore, finalGrade, gradeExplanation, updatedAt]
    AiResearch:
      type: object
      properties:
        symbol:
          type: string
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        opportunityLevel:
          type: string
          enum: [low, moderate, high, very_high]
        generatedAt:
          type: string
      required: [symbol, summary, strengths, weaknesses, risks, opportunityLevel, generatedAt]
    Narrative:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        iconKey:
          type: string
        momentumScore:
          type: integer
        perf24h:
          type: number
        perf7d:
          type: number
        tokenCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, iconKey, momentumScore, perf24h, perf7d, tokenCount, createdAt]
...
    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        source:
          type: string
        url:
          type: string
        summary:
          type: ["string", "null"]
        sentiment:
          type: string
          enum: [bullish, bearish, neutral]
        relatedSymbols:
          type: array
          items:
            type: string
        publishedAt:
          type: string
      required: [id, title, source, url, sentiment, relatedSymbols, publishedAt]